requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0
pyyaml>=6.0.1
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0
websockets>=12.0
aiofiles>=23.0.0
httpx>=0.25.0